move StatisticsAccumulator to mirgecom.utils
