+++ mirgecom/utils.py
@@ -27,6 +27,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
+__doc__ = """
+.. autoclass:: StatisticsAccumulator
+.. autofunction:: asdict_shallow
+"""
+
 
 def asdict_shallow(dc_instance):
     """Convert a dataclass into a dict.
@@ -38,3 +43,71 @@ def asdict_shallow(dc_instance):
     from dataclasses import fields
     return {attr.name: getattr(dc_instance, attr.name)
             for attr in fields(dc_instance)}
+
+
+class StatisticsAccumulator:
+    """Class that provides statistical functions for multiple values.
+
+    .. automethod:: __init__
+    .. automethod:: add_value
+    .. automethod:: sum
+    .. automethod:: mean
+    .. automethod:: max
+    .. automethod:: min
+    .. autoattribute:: num_values
+    """
+
+    def __init__(self, scale_factor: float = 1):
+        """Initialize an empty StatisticsAccumulator object.
+
+        Parameters
+        ----------
+        scale_factor
+            Scale returned statistics by this factor.
+        """
+        self.num_values = 0
+        """Number of values stored in the StatisticsAccumulator."""
+
+        self._sum = 0
+        self._min = None
+        self._max = None
+        self.scale_factor = scale_factor
+
+    def add_value(self, v: float):
+        """Add a new value to the statistics."""
+        if v is None:
+            return
+        self.num_values += 1
+        self._sum += v
+        if self._min is None or v < self._min:
+            self._min = v
+        if self._max is None or v > self._max:
+            self._max = v
+
+    def sum(self):
+        """Return the sum of added values."""
+        if self.num_values == 0:
+            return None
+
+        return self._sum / self.scale_factor
+
+    def mean(self):
+        """Return the mean of added values."""
+        if self.num_values == 0:
+            return None
+
+        return self._sum / self.num_values / self.scale_factor
+
+    def max(self):
+        """Return the max of added values."""
+        if self.num_values == 0:
+            return None
+
+        return self._max / self.scale_factor
+
+    def min(self):
+        """Return the min of added values."""
+        if self.num_values == 0:
+            return None
+
+        return self._min / self.scale_factor
